enable --interactive for 'sink' and remove --all flag
because now `sink -a` can be done by `collect -i tomorrow`
